requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-extensions==3.2.0
python-dotenv==0.20.0
gunicorn==20.1.0